requirements.txt
pydantic==2.7.2
parsel==1.9.1
requests==2.32.3
Scrapy==2.11.2
itemadapter==0.9.0
discord-webhook==1.3.0

python-dotenv==1.0.1